100-bb-Present-dispatch-edited-nowork
The best attempted were made to edit the data to attain results using RENPASSG!S. However, the following error message and information was provided by RENPASSG!S. 

Traceback (most recent call last):
  File "/home/g61/.virtualenvs/Phdstudy1/bin/renpass", line 6, in <module>
    exec(compile(open(__file__).read(), __file__, 'exec'))
  File "/home/g61/renpass/bin/renpass", line 14, in <module>
    renpass.cli.main(**arguments)
  File "/home/g61/renpass/renpass/cli.py", line 242, in main
    es = create_energysystem(arguments['DATAPACKAGE'], **arguments)
  File "/home/g61/renpass/renpass/cli.py", line 81, in create_energysystem
    typemap=options.typemap)
  File "/home/g61/oemof/oemof/tools/datapackage.py", line 140, in deserialize_energy_system
    (cast_error_msg).format(r.name))
tableschema.exceptions.CastError: Metadata structure of resource `volatile` does not match data structure. Check the column names, types and their order.

The best attempts were made to edit the data to ensure that the metadata and the data in the volatile.csv were matching, however, it does not work!!!! This is the latest attempt!